# pkg/server/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Peace Breaker Bot API
  version: "1.0"
  description: |
    HTTP API for controlling the Discord bot's playback and for reading
    from myinstants.com. This is the backend for the companion Electron/React
    UI in `peace-breaker-bot-desktop`.

    ## Versioning

    This document describes `v1`, rooted at `/api/v1`. `GET /api/docs`
    renders a human-readable version of it (and, once a v2 exists, a
    switcher between the two); it is deliberately not itself versioned,
    since it documents every version.

    ## Error labels

    `label` is a stable machine-readable code — match on it, not on
    `message` or the HTTP status. `message` defaults to English and
    negotiates via a standard `Accept-Language` header (currently `en-US`
    and `pt-BR` are supported; anything else falls back to English). The
    companion UI ignores `message` entirely and translates `label` itself,
    so it never sends this header. Envelopes distinguish success from error
    by the presence of `data` versus `label`, independent of status code —
    but the status code is now also meaningful and reflects each case below.
servers:
  - url: http://localhost:9001
    description: Default local bind address (server.address / --server-address / SERVER_ADDRESS)
security: []


paths:
  /api/v1/bot/play:
    post:
      operationId: botPlay
      summary: Play a clip through the bot
      description: |
        Resolves `url` to a cached/downloaded clip and starts playing it on
        the bot's current voice connection, stopping whatever is already
        playing first (a bot can only play one clip at a time). Blocks until
        playback ends naturally or is interrupted by `POST /api/v1/bot/stop`,
        then returns which one happened.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotPlayRequest"
      responses:
        "200":
          description: Playback finished or was interrupted; see `data.exitReason`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotPlaySuccessEnvelope"
              examples:
                playedToTheEnd:
                  summary: Playback finished on its own
                  value: { data: { exitReason: end } }
                stopped:
                  summary: Interrupted by POST /api/v1/bot/stop or a newer /api/v1/bot/play
                  value: { data: { exitReason: stop } }
        "400":
          description: The body wasn't valid JSON, or `url` is missing/malformed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                invalidBody:
                  summary: Body was not valid JSON
                  value: { label: invalid_body, message: "Invalid request" }
                invalidUrl:
                  summary: url is missing/malformed
                  value: { label: invalid_url, message: "That URL isn't valid" }
        "404":
          description: The clip could not be resolved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                instantNotFound:
                  value: { label: instant_not_found, message: "That instant couldn't be found" }
        "422":
          description: The downloaded file did not sniff as mp3.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                unsupportedAudioFormat:
                  value: { label: unsuported_audio_format, message: "That instant isn't an audio format we can play" }
        "500":
          description: Anything else.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                unknownError:
                  value: { label: unknown_error, message: "Something went wrong. Try again in a moment" }

  /api/v1/bot/stop:
    post:
      operationId: botStop
      summary: Stop whatever the bot is currently playing
      description: |
        Signals the in-flight `POST /api/v1/bot/play` (if any) to stop; that
        request then returns with `exitReason: "stop"`. Safe to call when
        nothing is playing (idempotent).
      responses:
        "200":
          description: Empty body; no Content-Type is set. Always this status.

  /api/v1/instants/{url}/content:
    get:
      operationId: getInstantContent
      summary: Fetch a clip as a base64 data URI
      description: |
        Downloads/caches the clip at `url` (same cache `POST /api/v1/bot/play`
        uses) and returns it inline, without touching the bot or its voice
        connection. Meant for local preview in the UI. `url` is the clip's
        full myinstants.com URL, percent-encoded into this single path
        segment (e.g. `:` becomes `%3A`, `/` becomes `%2F`).
      parameters:
        - name: url
          in: path
          required: true
          schema:
            type: string
            format: uri
          description: The myinstants.com clip URL to fetch, percent-encoded.
      responses:
        "200":
          description: The clip resolved, or it didn't — see `data.exists`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstantContentSuccessEnvelope"
              examples:
                found:
                  summary: Clip resolved and inlined
                  value: { data: { exists: true, content: "data:audio/mp3;base64,//uQx..." } }
                notFound:
                  summary: Clip could not be resolved
                  value: { data: { exists: false } }
        "400":
          description: The path segment didn't decode to a usable URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                invalidUrl:
                  value: { label: invalid_url, message: "That URL isn't valid" }
        "500":
          description: Anything else.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                unknownError:
                  value: { label: unknown_error, message: "Something went wrong. Try again in a moment" }

  /api/v1/instants:
    get:
      operationId: listInstants
      summary: List or search myinstants.com clips
      description: |
        Either browses a region's index (no `search`) or searches by name
        (`search`, which ignores `region`). `pages` is inferred from how
        many instants came back on this page, since myinstants.com's
        listing pages no longer carry a pager: a full page (36 instants)
        means there may be another, a short page means this is the last
        one, and an empty page means the caller paged past the end — that
        last case is normalized into the same collection shape as a normal
        result (`instants: []` plus the last valid `pages`), never a bare
        empty array.
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: 1-indexed page number. Anything non-positive or non-numeric is treated as 1.
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search term. When set, region is ignored.
        - name: region
          in: query
          required: false
          schema:
            type: string
            pattern: "^[a-z]{2}$"
            default: us
          description: Two-letter region code, case-insensitive. Ignored when search is set.
      responses:
        "200":
          description: One or more matches, or a search paged past its end.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstantListSuccessEnvelope"
              examples:
                results:
                  summary: One or more matches
                  value:
                    data:
                      instants:
                        - { name: "Bruh", url: "https://www.myinstants.com/media/sounds/bruh.mp3" }
                      pages: 2
                pastLastPage:
                  summary: A search paged past its end (upstream answered 404)
                  value: { data: { instants: [], pages: 4 } }
        "400":
          description: region did not match ^[a-z]{2}$.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                invalidRegion:
                  value: { label: invalid_region, message: "That region code isn't valid" }
        "500":
          description: The scrape found a different number of names and play links (markup likely changed), or anything else unexpected.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                nameLinkNotMatched:
                  value: { label: name_link_not_matched, message: "myinstants.com's results couldn't be read — their page layout may have changed" }
                unknownError:
                  value: { label: unknown_error, message: "Something went wrong. Try again in a moment" }
        "502":
          description: myinstants.com could not be reached, or answered with a bad status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                httpRequest:
                  summary: Could not reach myinstants.com
                  value: { label: http_request, message: "Couldn't reach myinstants.com" }
                badHttpStatus:
                  summary: myinstants.com answered a non-200/404 status
                  value: { label: bad_http_status, message: "myinstants.com answered with an error" }

  /api/v1/openapi.yaml:
    get:
      operationId: getOpenApiSpecV1
      summary: This document
      responses:
        "200":
          description: The v1 OpenAPI document, as YAML.

  /api/docs:
    get:
      operationId: getDocs
      summary: Human-readable rendering of the API's OpenAPI documents
      description: >-
        Not versioned — one docs page for every API version, rendered via
        Redoc with a version switcher above the spec (currently just `v1`).
      responses:
        "200":
          description: An HTML page rendering the available OpenAPI documents.

components:
  schemas:
    Envelope:
      type: object
      description: The shape every response uses, success or error.
      properties:
        label:
          type: string
          description: Present on error responses; a stable machine-readable code.
        message:
          type: string
          description: >-
            Present on error responses; a human-readable message, in
            English by default (see the Error labels section above for
            negotiating a different one via Accept-Language).
        data:
          description: Present on success responses; shape depends on the route.

    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [label, message]

    BotPlayRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: The myinstants.com clip URL to play.

    BotPlaySuccessEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [data]
          properties:
            data:
              type: object
              required: [exitReason]
              properties:
                exitReason:
                  type: string
                  enum: [end, stop]
                  description: end = playback finished naturally; stop = interrupted.

    InstantContentSuccessEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [data]
          properties:
            data:
              type: object
              required: [exists]
              properties:
                exists:
                  type: boolean
                content:
                  type: string
                  description: Only present when exists is true. A base64 data URI, e.g. "data:audio/mp3;base64,...".

    InstantButton:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
          format: uri

    InstantListSuccessEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [data]
          properties:
            data:
              type: object
              required: [instants, pages]
              properties:
                instants:
                  type: array
                  items:
                    $ref: "#/components/schemas/InstantButton"
                  description: Always present, even when empty (e.g. a search paged past its end).
                pages:
                  type: integer